+++ CHANGELOG.md
@@ -21,3 +21,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - User Service
 - Project Role Constraints Aspect
 - Extract path roles map for Security Configuration
+- Project State Constraints Aspect